Struct safe_client::client::response_getter::ResponseGetter

pub struct ResponseGetter {
    // some fields omitted
}

ResponseGetter is a lazy evaluated response getter. It will fetch either from local cache or wait for the MessageQueue to notify it of the incoming response from the network.

Methods

impl ResponseGetter

fn new(notifier: Option<ResponseNotifier>, message_queue: Arc<Mutex<MessageQueue>>, requested_type: DataRequest) -> ResponseGetter

Create a new instance of ResponseGetter

fn get(&self) -> Result<Data, ClientError>

Get either from local cache or (if not available there) get it when it comes from the network as informed by MessageQueue. This is blocking.