Module safe_client::structured_data_operations

Helper functions to handle StructuredData related operations

Modules

unversioned
Unversioned-Structured Data

versioned
Versioned-Structured Data

Enums

DataFitResult
Inform about data fitting or not into given StructuredData

Functions

check_if_data_can_fit_in_structured_data
Check if it is possible to fit the given data into the given StructuredData

get_approximate_space_for_data
Calculates approximate space available for data. Calculates the worst case scenario in which all owners must sign this StructuredData.